+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -60,5 +60,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-13T22:21:46.762Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-20T11:35:16.103Z")
     r"""The date and time the resource was created"""